added access management added user activity logging added security headings added secure randomness finished the project

+++ lottery/views.py
@@ -2,24 +2,26 @@
 import logging
 
 from flask import Blueprint, render_template, request, flash
-from flask_login import login_required
-from app import db
+from flask_login import login_required, current_user
+from app import db, requires_roles
 from models import Draw, User
 
 # CONFIG
 lottery_blueprint = Blueprint('lottery', __name__, template_folder='templates')
 
-user = User.query.first()
-draw_key = user.draw_key
 
 # VIEWS
 # view lottery page
 @lottery_blueprint.route('/lottery')
+@login_required
+@requires_roles('user')
 def lottery():
     return render_template('lottery.html')
 
 
 @lottery_blueprint.route('/add_draw', methods=['POST'])
+@login_required
+@requires_roles('user')
 def add_draw():
     submitted_draw = ''
     for i in range(6):
@@ -27,7 +29,8 @@ def add_draw():
     submitted_draw.strip()
 
     # create a new draw with the form data.
-    new_draw = Draw(user_id=1, draw=submitted_draw, win=False, round=0, draw_key=draw_key)  # TODO: update user_id [user_id=1 is a placeholder]
+
+    new_draw = Draw(user_id=current_user.id, draw=submitted_draw, win=False, round=0, draw_key=current_user.draw_key)
 
     # add the new draw to the database
     db.session.add(new_draw)
@@ -40,11 +43,13 @@ def add_draw():
 
 # view all draws that have not been played
 @lottery_blueprint.route('/view_draws', methods=['POST'])
+@login_required
+@requires_roles('user')
 def view_draws():
     # get all draws that have not been played [played=0]
-    playable_draws = Draw.query.filter_by(played=False).all() # TODO: filter playable draws for current user
+    playable_draws = Draw.query.filter_by(played=False, user_id=current_user.id).all()
     for p in playable_draws:
-        p.view_draw(draw_key)
+        p.view_draw(current_user.draw_key)
 
     # if playable draws exist
     if len(playable_draws) != 0:
@@ -57,11 +62,13 @@ def view_draws():
 
 # view lottery results
 @lottery_blueprint.route('/check_draws', methods=['POST'])
+@login_required
+@requires_roles('user')
 def check_draws():
     # get played draws
-    played_draws = Draw.query.filter_by(played=True).all()  # TODO: filter played draws for current user
+    played_draws = Draw.query.filter_by(played=True, user_id=current_user.id).all()
     for p in played_draws:
-        p.view_draw(draw_key)
+        p.view_draw(current_user.draw_key)
 
     # if played draws exist
     if len(played_draws) != 0:
@@ -75,12 +82,13 @@ def check_draws():
 
 # delete all played draws
 @lottery_blueprint.route('/play_again', methods=['POST'])
+@login_required
+@requires_roles('user')
 def play_again():
-    delete_played = Draw.__table__.delete().where(Draw.played)  # TODO: delete played draws for current user only
+    delete_played = Draw.__table__.delete().where(Draw.played & Draw.user_id == current_user.id)
+
     db.session.execute(delete_played)
     db.session.commit()
 
     flash("All played draws deleted.")
     return lottery()
-
-
